Added sharing of Events

+++ backend/routes/web.php
@@ -4,6 +4,8 @@ use App\Http\Controllers\EventController;
 use App\Http\Controllers\EventProductLinkController;
 use App\Http\Controllers\ProductController;
 use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\ShareController;
+use App\Http\Middleware\EventOwnerOrShareOnly;
 use App\Models\Event;
 use Illuminate\Support\Facades\Route;
 
@@ -23,15 +25,19 @@ Route::get('/', function () {
 });
 
 Route::middleware('auth')->group(function () {
-    Route::resource('events', EventController::class);
-    Route::put("events/{event}/status", [EventController::class, 'toggleStatus'])->name("events.status.toggle");
-    Route::post("events/{event}/report", [EventController::class, 'generateReport'])->name("events.report.create");
+    Route::resource('events', EventController::class)->middleware(EventOwnerOrShareOnly::class);
+    Route::put("events/{event}/status", [EventController::class, 'toggleStatus'])->name("events.status.toggle")->middleware(EventOwnerOrShareOnly::class);
+    Route::post("events/{event}/report", [EventController::class, 'generateReport'])->name("events.report.create")->middleware(EventOwnerOrShareOnly::class);
     Route::resource('products', ProductController::class);
 
-    Route::get('events/{event}/products/add', [EventProductLinkController::class, 'create'])->name("events.products.add");
-    Route::post('events/{event}/products/add', [EventProductLinkController::class, 'store'])->name("events.products.store");
-    Route::patch('events/{event}/products/', [EventProductLinkController::class, 'update'])->name("events.products.update");
-    Route::delete('events/{event}/products/', [EventProductLinkController::class, 'destroy'])->name("events.products.destroy");
+    Route::get('events/{event}/products/add', [EventProductLinkController::class, 'create'])->name("events.products.add")->middleware(EventOwnerOrShareOnly::class);
+    Route::post('events/{event}/products/add', [EventProductLinkController::class, 'store'])->name("events.products.store")->middleware(EventOwnerOrShareOnly::class);
+    Route::patch('events/{event}/products/', [EventProductLinkController::class, 'update'])->name("events.products.update")->middleware(EventOwnerOrShareOnly::class);
+    Route::delete('events/{event}/products/', [EventProductLinkController::class, 'destroy'])->name("events.products.destroy")->middleware(EventOwnerOrShareOnly::class);
+
+    Route::get('events/{event}/shares', [ShareController::class, 'index'])->name("events.shares.index")->middleware(EventOwnerOrShareOnly::class);
+    Route::post('events/{event}/shares/store', [ShareController::class, 'store'])->name("events.shares.store")->middleware(EventOwnerOrShareOnly::class);
+    Route::delete('events/{event}/shares/{share}/destroy', [ShareController::class, 'destroy'])->name("events.shares.destroy")->middleware(EventOwnerOrShareOnly::class);
 });
